backend/requirements.txt
fastapi
uvicorn
numpy
matplotlib
google-generativeai
python-multipart
pillow
scipy
pydantic
